Latest

+++ controller/contacts.controller.js
@@ -2,41 +2,99 @@ import Contact from "../models/contacts.models.js"
 import mongoose from "mongoose"
                           
 export  const getContacts = async (req,res) => { 
+
+try {
     const contacts = await Contact.find(req.params.id )
-    res.render('home',{contacts})
+    return res.render('home',{contacts})
+
+}catch(error){
+    return res.render('500',{message: error})
+    }
+
+
+    
     }
 
 export  const getContact = async (req,res) => { 
 
-    if (!mongoose.Types.ObjectId.isValid(req.params.id)){
-        res.render('404',{message:"Contact Not Found! "})
+    if(!mongoose.Types.ObjectId.isValid(req.params.id)){
+       return  res.render('404',{message:"Invalid Ids! "})
     }
 
-    const contact = await Contact.findById(req.params.id )
-    res.render('show-contact',{contact})
+    try{
+        const contact = await Contact.findById(req.params.id )
+    if(!contact) return res.render('404',{message:"Conatct not Found"})    
+    return res.render('show-contact',{contact})}
+    catch(error){
+    return res.render('500',{message: error})
+    }
+
+    
     }
 
 export const addContactPage = (req,res) =>{ 
-    res.render('add-contact')
+    return res.render('add-contact')
 }
 
  export const addContact = async(req,res)=>{
-    await Contact.create(req.body) 
-    res.redirect("/")}
 
+     try{
+  await Contact.create(req.body) 
+    return res.redirect("/")}
+   catch(error){
+ return res.render('500',{message: error})
+
+    }
+}
 
 export const updateContactPage = async(req,res) => { 
-    const contact = await Contact.findById(req.params.id)
-    res.render('update-contact',{contact})   
+    if(!mongoose.Types.ObjectId.isValid(req.params.id)){
+       return  res.render('404',{message:"Invalid Id! "})
+    }
+    try{
+ const contact = await Contact.findById(req.params.id)
+ if(!contact) return res.render('404',{message:"Conatct not Found"})   
+ return res.render('update-contact',{contact})   
+
+    }catch(error){
+ return res.render('500',{message: error})
+
+    }
+   
 }
 
 export const updateContact = async(req,res)=>{
-const {first_name, last_name, email, phone, address} = req.body
-await Contact.findByIdAndUpdate(req.params.id, {first_name, last_name, email, phone, address})
-res.redirect("/")
+    if(!mongoose.Types.ObjectId.isValid(req.params.id)){
+        return res.render('404',{message:"Invalid Id! "})
+    }
+    try{
+   const contact = await Contact.findByIdAndUpdate(req.params.id,req.body)
+ if(!contact) return res.render('404',{message:"Conatct not Found"})   
+    return res.redirect("/")
+
+    }catch(error){
+ return res.render('500',{message: error})
+
+    }
+
+// const {first_name, last_name, email, phone, address} = req.body
+// await Contact.findByIdAndUpdate(req.params.id, {first_name, last_name, email, phone, address})
+// res.redirect("/")
  }
 
 export const deleteContact = async(req,res)=>{ 
-await Contact.findByIdAndDelete(req.params.id)
-res.redirect("/")   
+
+if(!mongoose.Types.ObjectId.isValid(req.params.id)){
+        return res.render('404',{message:"Invalid Id! "})
+    }
+
+        try{
+const contact = await Contact.findByIdAndDelete(req.params.id)
+ if(!contact) return res.render('404',{message:"Conatct not Found"})   
+    return res.redirect("/")
+
+    }catch(error){
+ return res.render('500',{message: error})
+
+    }
 }
